+++ NEWS.md
@@ -5,6 +5,7 @@
 
 - Now that `@__DIR__` is provided, do not `cd()` into the test directory.
 - Allow running tests in the context of a specific module.
+- Added an option to suppress any output except for error reports.
 
 
 ## v0.4.1